requirements.txt
python-yaml
python-cairo